update v1 and change int->double of SizeProduct

+++ src/main/java/com/capstone/pod/dto/sizeproduct/AddSizeProductDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 public class AddSizeProductDto {
     private String size;
-    private int width;
-    private int height;
+    private double width;
+    private double height;
     private String productId;
 }
