+++ src/main/java/com/capstone/pod/entities/SizeProduct.java
@@ -21,8 +21,8 @@ public class SizeProduct {
     private String id;
     private String size;
 
-    private int width;
-    private int height;
+    private double width;
+    private double height;
 
     @ManyToOne
     private Product product;
